feat(game): improve recommended games UI
- Rename Related Games to Recommended Games to more accurately reflect its purpose.
- Truncate long game titles and display their full names in tooltips for a cleaner browsing experience.
