+++ src/renderer/locales/zh-CN/game.json
@@ -218,11 +218,11 @@
         "tags": "标签",
         "basicInfo": "基本信息",
         "extraInfo": "附加信息",
-        "relatedGames": "相关游戏",
+        "recommendedGames": "推荐游戏",
         "relatedSites": "相关网站"
       },
-      "relatedGames": {
-        "empty": "暂无相关游戏"
+      "recommendedGames": {
+        "empty": "暂无推荐游戏"
       },
       "description": {
         "empty": "暂无简介",
