+++ src/renderer/locales/zh-TW/game.json
@@ -218,11 +218,11 @@
         "tags": "標籤",
         "basicInfo": "基本資訊",
         "extraInfo": "附加資訊",
-        "relatedGames": "相關遊戲",
+        "recommendedGames": "推薦遊戲",
         "relatedSites": "相關網站"
       },
-      "relatedGames": {
-        "empty": "暫無相關遊戲"
+      "recommendedGames": {
+        "empty": "暫無推薦遊戲"
       },
       "description": {
         "empty": "暫無簡介",
